requirements.txt
flask
openai
watchdog
python-dotenv
